Graph code extracted, sparkline graph added
